+++ crates/action-runner/src/subscribers/local_cache.rs
@@ -1,4 +1,4 @@
-use crate::events::Event;
+use crate::emitter::Event;
 use moon_contract::EventFlow;
 use moon_error::MoonError;
 use moon_utils::path;
